app/requirements.txt
fastapi[standard]
uvicorn[standard]
paho-mqtt
stmpy